Moved utils to utils.
